Correct the npm name used to output the minimal package.json on install
Before it was insufficiently decoding the `npm-gen-my_module` name into
`my_module`. Which is fine if the module looks like `my_module`. However
if the module is namespaced `@vue/cli` then this gets encoded into
`npm-gen-at_vue-cli` then decoded again into `at_vue-cli` which doesn't
exist in either the shrinkwrap.json or on npm.

This change makes use of the shrinkwrap format that is generated where
the source module is actually the first (and only) dependency of the
generated wrapper module, so we can use that to work out the real npm
name instead of trying to decode it from the wrapper module name.